Separate context building from rendering
